requirements.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
deepgram-sdk==3.2.0
chromadb
sentence-transformers
ollama
websockets
httpx
pydantic
torch
torchvision
torchaudio

sqlalchemy

torch 
torchaudio
pytz 
datetime
PyJWT

python-multipart    